Cleaned up logic for when player isn't running.
Previously, if you requested a certain player and none were running, it
would tell you "hey that one's not running but here's a list of players
that are running", but then that list would be empty. Now it will just
tell you no players are running.

I think it also handles oddball non-compliant players better too, as it
will die with an informative backtrace instead of covering it up and
telling you that the player simply isn't there.

I also moved all of the printing to stderr and exiting code to the main
block. The OO part of the code now just throws exceptions.

I also added some relevant tests.